requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.14
fastapi-cli==0.0.7
frozenlist==1.7.0
gotrue==2.12.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.5.1
orjson==3.10.18
packaging==25.0
pluggy==1.6.0
postgrest==1.1.1
propcache==0.3.2
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyTelegramBotAPI==4.27.0
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.5.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
